Update skeleton for 2020 petition site

+++ app/models/backdoor.model.js
@@ -2,20 +2,17 @@ const db = require('../../config/db');
 const fs = require('mz/fs');
 
 const photoDirectory = './storage/photos/';
+const defaultPhotoDirectory = './storage/default/';
 
-exports.resetDB = async function () {
+exports.resetDb = async function () {
     let promises = [];
 
     const sql = await fs.readFile('app/resources/create_database.sql', 'utf8');
-    promises.push(db.getPool().query(sql));
+    promises.push(db.getPool().query(sql));  // sync call to recreate DB
 
-    if (await fs.exists(photoDirectory)) {
-        const files = await fs.readdir(photoDirectory);
-        for (const file of files) {
-            if (file !== 'default.png') {
-                promises.push(fs.unlink(photoDirectory + file));
-            }
-        }
+    const files = await fs.readdir(photoDirectory);
+    for (const file of files) {
+        if (file !== '.gitkeep') promises.push(fs.unlink(photoDirectory + file));  // sync call to delete photo
     }
 
     return Promise.all(promises);  // async wait for DB recreation and photos to be deleted
@@ -30,6 +27,10 @@ exports.loadData = async function () {
         console.log(err.sql);
         throw err;
     }
+
+    const defaultPhotos = await fs.readdir(defaultPhotoDirectory);
+    const promises = defaultPhotos.map(file => fs.copyFile(defaultPhotoDirectory + file, photoDirectory + file));
+    return Promise.all(promises);
 };
 
 /**
@@ -38,12 +39,13 @@ exports.loadData = async function () {
  * @returns {Promise<void>}
  */
 async function populateDefaultUsers() {
-    const createSQL = 'INSERT INTO User (username, email, given_name, family_name, password) VALUES ?';
-    let { properties, usersData } = require('../resources/default_users');
+    const createSQL = `INSERT INTO User (name, email, password, city, country, photo_filename)
+                       VALUES ?`;
+    let {properties, usersData} = require('../resources/default_users');
 
     // Shallow copy all the user arrays within the main data array
     // Ensures that the user arrays with hashed passwords won't persist across multiple calls to this function
-    usersData = usersData.map(user => ([ ...user ]));
+    usersData = usersData.map(user => ([...user]));
 
     const passwordIndex = properties.indexOf('password');
     await Promise.all(usersData.map(user => changePasswordToHash(user, passwordIndex)));
@@ -65,7 +67,8 @@ async function changePasswordToHash(user, passwordIndex) {
 
 exports.executeSql = async function (sql) {
     try {
-        return await db.getPool().query(sql);
+        const [rows] = await db.getPool().query(sql);
+        return rows;
     } catch (err) {
         console.log(err.sql);
         throw err;
